Windows: Move the graphics context creation to main.cpp
